sqlscripts renamed to sqlqueries, and more verbose package comment added
