+++ pkg/sqlqueries/sql_queries.go
@@ -1,0 +1,4 @@
+// Package sqlqueries holds references to relevant SQL queries
+// it exports variables to hold references to relevant queries, organized by relevant entity
+// it's sole responsibility is embedding sql queries for use by other modules which use these queries to interact with the database
+package sqlqueries
