+++ pkg/sqlqueries/trb_request_form.go
@@ -1,4 +1,4 @@
-package sqlscripts
+package sqlqueries
 
 import _ "embed"
