--- pkg/sqlutils/sqlscripts/sql_scripts.go
@@ -1,2 +1,0 @@
-// Package sqlscripts holds references to relevant SQL queries
-package sqlscripts
