+++ pkg/storage/trb_request.go
@@ -11,7 +11,7 @@ import (
 	"github.com/cmsgov/easi-app/pkg/appcontext"
 	"github.com/cmsgov/easi-app/pkg/apperrors"
 	"github.com/cmsgov/easi-app/pkg/models"
-	"github.com/cmsgov/easi-app/pkg/sqlutils/sqlscripts"
+	"github.com/cmsgov/easi-app/pkg/sqlqueries"
 
 	_ "embed"
 )
@@ -25,7 +25,7 @@ func (s *Store) CreateTRBRequest(ctx context.Context, trb *models.TRBRequest) (*
 		trb.ID = uuid.New()
 	}
 
-	stmt, err := tx.PrepareNamed(sqlscripts.TRBRequest.Create)
+	stmt, err := tx.PrepareNamed(sqlqueries.TRBRequest.Create)
 	if err != nil {
 		appcontext.ZLogger(ctx).Error(
 			fmt.Sprintf("Failed to create trb request with error %s", err),
@@ -52,7 +52,7 @@ func (s *Store) CreateTRBRequest(ctx context.Context, trb *models.TRBRequest) (*
 	form.ID = uuid.New()
 	form.CreatedBy = retTRB.CreatedBy
 
-	stmt, err = tx.PrepareNamed(sqlscripts.TRBRequestForm.Create)
+	stmt, err = tx.PrepareNamed(sqlqueries.TRBRequestForm.Create)
 
 	if err != nil {
 		appcontext.ZLogger(ctx).Error(
@@ -84,7 +84,7 @@ func (s *Store) CreateTRBRequest(ctx context.Context, trb *models.TRBRequest) (*
 func (s *Store) GetTRBRequestByID(ctx context.Context, id uuid.UUID) (*models.TRBRequest, error) {
 
 	trb := models.TRBRequest{}
-	stmt, err := s.db.PrepareNamed(sqlscripts.TRBRequest.GetByID)
+	stmt, err := s.db.PrepareNamed(sqlqueries.TRBRequest.GetByID)
 	if err != nil {
 		appcontext.ZLogger(ctx).Error(
 			"Failed to fetch TRB request",
@@ -113,7 +113,7 @@ func (s *Store) GetTRBRequestByID(ctx context.Context, id uuid.UUID) (*models.TR
 
 // UpdateTRBRequest returns an TRBRequest from the db for a given id
 func (s *Store) UpdateTRBRequest(ctx context.Context, trb *models.TRBRequest) (*models.TRBRequest, error) {
-	stmt, err := s.db.PrepareNamed(sqlscripts.TRBRequest.Update)
+	stmt, err := s.db.PrepareNamed(sqlqueries.TRBRequest.Update)
 	if err != nil {
 		appcontext.ZLogger(ctx).Error(
 			fmt.Sprintf("Failed to update TRB request %s", err),
@@ -143,7 +143,7 @@ func (s *Store) UpdateTRBRequest(ctx context.Context, trb *models.TRBRequest) (*
 func (s *Store) GetTRBRequests(ctx context.Context, archived bool) ([]*models.TRBRequest, error) {
 	trbRequests := []*models.TRBRequest{}
 
-	stmt, err := s.db.PrepareNamed(sqlscripts.TRBRequest.CollectionGet)
+	stmt, err := s.db.PrepareNamed(sqlqueries.TRBRequest.CollectionGet)
 	if err != nil {
 		appcontext.ZLogger(ctx).Error(
 			"Failed to fetch trb requests",
@@ -173,7 +173,7 @@ func (s *Store) GetTRBRequests(ctx context.Context, archived bool) ([]*models.TR
 func (s *Store) GetMyTRBRequests(ctx context.Context, archived bool) ([]*models.TRBRequest, error) {
 	trbRequests := []*models.TRBRequest{}
 
-	stmt, err := s.db.PrepareNamed(sqlscripts.TRBRequest.CollectionGetByUserAndArchivedState)
+	stmt, err := s.db.PrepareNamed(sqlqueries.TRBRequest.CollectionGetByUserAndArchivedState)
 	if err != nil {
 		appcontext.ZLogger(ctx).Error(
 			"Failed to fetch user's trb requests",
